Check for spaces in filename

This function checks if the provided filename contains any spaces.

Usage

check_filename_spaces(filename)

Arguments

filename

A character string of the filename to check

Value

A list containing:

check

Character string with the stage that the check belongs to.

result

Character string, either "PASS", "FAIL" or "ADVISORY".

message

Character string with feedback about the check.

Examples

check_filename_spaces("datafile.csv")
#> $check
#> [1] "check_filename_spaces"
#> 
#> $result
#> [1] "PASS"
#> 
#> $message
#> [1] "'datafile.csv' does not have spaces in the filename."
#> 
check_filename_spaces("data file.meta.csv")
#> $check
#> [1] "check_filename_spaces"
#> 
#> $result
#> [1] "FAIL"
#> 
#> $message
#> [1] "There are spaces that need removing in 'data file.meta.csv' (filename)."
#>