Check for special characters in filename

This function checks if the provided filename contains any special characters.

Usage

check_filename_special(filename, custom_name = NULL, output = "console")

Arguments

filename

A character string of the filename to check

custom_name

An optional character string to customize the test name in the output

output

"table", "error-only", or "console". Default is "console"

Value

a console message with feedback from the check or a single row data frame if verbose = FALSE

See also

Examples

check_filename_special("datafile.csv")
#> ✔ 'datafile.csv' does not contain any special characters.
check_filename_special("data+file.meta.csv", output = "table")
#>                     check result
#> 1 check_filename__special   FAIL
#>                                                                                                                                                       message
#> 1 The following special characters need removing from data+file.meta.csv (filename): +. Filenames must only contain numbers, letters, hyphens or underscores.
#>   guidance_url
#> 1           NA
check_filename_special("datafile.meta.csv", custom_name = "meta")
#> ✔ 'datafile.meta.csv' does not contain any special characters.