Check filenames line up between data and metadata files

Making sure the data and metadata file follow the pattern of:

datafile.csv (data file)

datafile.meta.csv (metadata file)

Usage

check_filenames_match(datafilename, metafilename, output = "console")

Arguments

datafilename

A character string of the data filename to check

metafilename

A character string of the metadata filename to check

output

"table", "error-only", or "console". Default is "console"

Value

a console message with feedback from the check or a single row data frame if verbose = FALSE

See also

Examples

check_filenames_match("datafile.csv", "datafile.meta.csv")
#> ✔ The names of the files follow the recommended naming convention.
check_filenames_match("datafile.csv", "metafile.csv", output = "table")
#>                   check result
#> 1 check_filenames_match   FAIL
#>                                                                                                                                                          message
#> 1 The filenames do not follow the recommended naming convention. Based on the given data filename, the metadata filename is expected to be 'datafile.meta.csv''.
#>   guidance_url
#> 1           NA