Generate a data file for testing

This function is designed to be used within generate_test_dfs to complement the meta file generated by generate_meta_file.

Usage

generate_data_file(
  years,
  pcon_names,
  pcon_codes,
  num_filters,
  num_indicators,
  verbose = FALSE
)

Arguments

years

A vector of 4 digit years

pcon_names

A vector of PCon names

pcon_codes

A vector of PCon codes

num_filters

The number of filters to generate, currently will give each filter 5 items

num_indicators

The number of indicators to generate, currently generates random numbers between 100 and 1000

verbose

Optional console messages highlighting the number of rows being created

Value

A data frame following the EES data standards

Details

Use the parameters to quickly create beefy files!