EES-ily generate some beefy test files

Mostly for ease of performance benchmarking. A bit crude, only does PCons for simplicity so far.

Usage

generate_test_dfs(
  years = 3000,
  pcon_names = "Sheffield Central",
  pcon_codes = "E14000919",
  num_filters = 1,
  num_indicators = 1,
  verbose = FALSE
)

Arguments

years

A vector of 4 digit years

pcon_names

A vector of PCon names

pcon_codes

A vector of PCon codes

num_filters

The number of filters to generate, currently will give each filter 5 items

num_indicators

The number of indicators to generate, currently generates random numbers between 100 and 1000

verbose

Optional console messages highlighting the number of rows being created

Value

A list containing two data frames, one data and associated meta

Examples

generate_test_dfs(2010:2015, "Sheffield Central", "E14000919", 2, 3)
#> $data
#>     time_period time_identifier           geographic_level country_name
#> 1          2010   Calendar year Parliamentary constituency      England
#> 2          2010   Calendar year Parliamentary constituency      England
#> 3          2010   Calendar year Parliamentary constituency      England
#> 4          2010   Calendar year Parliamentary constituency      England
#> 5          2010   Calendar year Parliamentary constituency      England
#> 6          2011   Calendar year Parliamentary constituency      England
#> 7          2011   Calendar year Parliamentary constituency      England
#> 8          2011   Calendar year Parliamentary constituency      England
#> 9          2011   Calendar year Parliamentary constituency      England
#> 10         2011   Calendar year Parliamentary constituency      England
#> 11         2012   Calendar year Parliamentary constituency      England
#> 12         2012   Calendar year Parliamentary constituency      England
#> 13         2012   Calendar year Parliamentary constituency      England
#> 14         2012   Calendar year Parliamentary constituency      England
#> 15         2012   Calendar year Parliamentary constituency      England
#> 16         2013   Calendar year Parliamentary constituency      England
#> 17         2013   Calendar year Parliamentary constituency      England
#> 18         2013   Calendar year Parliamentary constituency      England
#> 19         2013   Calendar year Parliamentary constituency      England
#> 20         2013   Calendar year Parliamentary constituency      England
#> 21         2014   Calendar year Parliamentary constituency      England
#> 22         2014   Calendar year Parliamentary constituency      England
#> 23         2014   Calendar year Parliamentary constituency      England
#> 24         2014   Calendar year Parliamentary constituency      England
#> 25         2014   Calendar year Parliamentary constituency      England
#> 26         2015   Calendar year Parliamentary constituency      England
#> 27         2015   Calendar year Parliamentary constituency      England
#> 28         2015   Calendar year Parliamentary constituency      England
#> 29         2015   Calendar year Parliamentary constituency      England
#> 30         2015   Calendar year Parliamentary constituency      England
#> 31         2010   Calendar year Parliamentary constituency      England
#> 32         2010   Calendar year Parliamentary constituency      England
#> 33         2010   Calendar year Parliamentary constituency      England
#> 34         2010   Calendar year Parliamentary constituency      England
#> 35         2010   Calendar year Parliamentary constituency      England
#> 36         2011   Calendar year Parliamentary constituency      England
#> 37         2011   Calendar year Parliamentary constituency      England
#> 38         2011   Calendar year Parliamentary constituency      England
#> 39         2011   Calendar year Parliamentary constituency      England
#> 40         2011   Calendar year Parliamentary constituency      England
#> 41         2012   Calendar year Parliamentary constituency      England
#> 42         2012   Calendar year Parliamentary constituency      England
#> 43         2012   Calendar year Parliamentary constituency      England
#> 44         2012   Calendar year Parliamentary constituency      England
#> 45         2012   Calendar year Parliamentary constituency      England
#> 46         2013   Calendar year Parliamentary constituency      England
#> 47         2013   Calendar year Parliamentary constituency      England
#> 48         2013   Calendar year Parliamentary constituency      England
#> 49         2013   Calendar year Parliamentary constituency      England
#> 50         2013   Calendar year Parliamentary constituency      England
#> 51         2014   Calendar year Parliamentary constituency      England
#> 52         2014   Calendar year Parliamentary constituency      England
#> 53         2014   Calendar year Parliamentary constituency      England
#> 54         2014   Calendar year Parliamentary constituency      England
#> 55         2014   Calendar year Parliamentary constituency      England
#> 56         2015   Calendar year Parliamentary constituency      England
#> 57         2015   Calendar year Parliamentary constituency      England
#> 58         2015   Calendar year Parliamentary constituency      England
#> 59         2015   Calendar year Parliamentary constituency      England
#> 60         2015   Calendar year Parliamentary constituency      England
#> 61         2010   Calendar year Parliamentary constituency      England
#> 62         2010   Calendar year Parliamentary constituency      England
#> 63         2010   Calendar year Parliamentary constituency      England
#> 64         2010   Calendar year Parliamentary constituency      England
#> 65         2010   Calendar year Parliamentary constituency      England
#> 66         2011   Calendar year Parliamentary constituency      England
#> 67         2011   Calendar year Parliamentary constituency      England
#> 68         2011   Calendar year Parliamentary constituency      England
#> 69         2011   Calendar year Parliamentary constituency      England
#> 70         2011   Calendar year Parliamentary constituency      England
#> 71         2012   Calendar year Parliamentary constituency      England
#> 72         2012   Calendar year Parliamentary constituency      England
#> 73         2012   Calendar year Parliamentary constituency      England
#> 74         2012   Calendar year Parliamentary constituency      England
#> 75         2012   Calendar year Parliamentary constituency      England
#> 76         2013   Calendar year Parliamentary constituency      England
#> 77         2013   Calendar year Parliamentary constituency      England
#> 78         2013   Calendar year Parliamentary constituency      England
#> 79         2013   Calendar year Parliamentary constituency      England
#> 80         2013   Calendar year Parliamentary constituency      England
#> 81         2014   Calendar year Parliamentary constituency      England
#> 82         2014   Calendar year Parliamentary constituency      England
#> 83         2014   Calendar year Parliamentary constituency      England
#> 84         2014   Calendar year Parliamentary constituency      England
#> 85         2014   Calendar year Parliamentary constituency      England
#> 86         2015   Calendar year Parliamentary constituency      England
#> 87         2015   Calendar year Parliamentary constituency      England
#> 88         2015   Calendar year Parliamentary constituency      England
#> 89         2015   Calendar year Parliamentary constituency      England
#> 90         2015   Calendar year Parliamentary constituency      England
#> 91         2010   Calendar year Parliamentary constituency      England
#> 92         2010   Calendar year Parliamentary constituency      England
#> 93         2010   Calendar year Parliamentary constituency      England
#> 94         2010   Calendar year Parliamentary constituency      England
#> 95         2010   Calendar year Parliamentary constituency      England
#> 96         2011   Calendar year Parliamentary constituency      England
#> 97         2011   Calendar year Parliamentary constituency      England
#> 98         2011   Calendar year Parliamentary constituency      England
#> 99         2011   Calendar year Parliamentary constituency      England
#> 100        2011   Calendar year Parliamentary constituency      England
#> 101        2012   Calendar year Parliamentary constituency      England
#> 102        2012   Calendar year Parliamentary constituency      England
#> 103        2012   Calendar year Parliamentary constituency      England
#> 104        2012   Calendar year Parliamentary constituency      England
#> 105        2012   Calendar year Parliamentary constituency      England
#> 106        2013   Calendar year Parliamentary constituency      England
#> 107        2013   Calendar year Parliamentary constituency      England
#> 108        2013   Calendar year Parliamentary constituency      England
#> 109        2013   Calendar year Parliamentary constituency      England
#> 110        2013   Calendar year Parliamentary constituency      England
#> 111        2014   Calendar year Parliamentary constituency      England
#> 112        2014   Calendar year Parliamentary constituency      England
#> 113        2014   Calendar year Parliamentary constituency      England
#> 114        2014   Calendar year Parliamentary constituency      England
#> 115        2014   Calendar year Parliamentary constituency      England
#> 116        2015   Calendar year Parliamentary constituency      England
#> 117        2015   Calendar year Parliamentary constituency      England
#> 118        2015   Calendar year Parliamentary constituency      England
#> 119        2015   Calendar year Parliamentary constituency      England
#> 120        2015   Calendar year Parliamentary constituency      England
#> 121        2010   Calendar year Parliamentary constituency      England
#> 122        2010   Calendar year Parliamentary constituency      England
#> 123        2010   Calendar year Parliamentary constituency      England
#> 124        2010   Calendar year Parliamentary constituency      England
#> 125        2010   Calendar year Parliamentary constituency      England
#> 126        2011   Calendar year Parliamentary constituency      England
#> 127        2011   Calendar year Parliamentary constituency      England
#> 128        2011   Calendar year Parliamentary constituency      England
#> 129        2011   Calendar year Parliamentary constituency      England
#> 130        2011   Calendar year Parliamentary constituency      England
#> 131        2012   Calendar year Parliamentary constituency      England
#> 132        2012   Calendar year Parliamentary constituency      England
#> 133        2012   Calendar year Parliamentary constituency      England
#> 134        2012   Calendar year Parliamentary constituency      England
#> 135        2012   Calendar year Parliamentary constituency      England
#> 136        2013   Calendar year Parliamentary constituency      England
#> 137        2013   Calendar year Parliamentary constituency      England
#> 138        2013   Calendar year Parliamentary constituency      England
#> 139        2013   Calendar year Parliamentary constituency      England
#> 140        2013   Calendar year Parliamentary constituency      England
#> 141        2014   Calendar year Parliamentary constituency      England
#> 142        2014   Calendar year Parliamentary constituency      England
#> 143        2014   Calendar year Parliamentary constituency      England
#> 144        2014   Calendar year Parliamentary constituency      England
#> 145        2014   Calendar year Parliamentary constituency      England
#> 146        2015   Calendar year Parliamentary constituency      England
#> 147        2015   Calendar year Parliamentary constituency      England
#> 148        2015   Calendar year Parliamentary constituency      England
#> 149        2015   Calendar year Parliamentary constituency      England
#> 150        2015   Calendar year Parliamentary constituency      England
#>     country_code         pcon_name pcon_code filter1  filter2 indicator1
#> 1      E92000001 Sheffield Central E14000919   Alpha   Alpha2        272
#> 2      E92000001 Sheffield Central E14000919   Bravo   Alpha2        506
#> 3      E92000001 Sheffield Central E14000919 Charlie   Alpha2        559
#> 4      E92000001 Sheffield Central E14000919   Delta   Alpha2        162
#> 5      E92000001 Sheffield Central E14000919    Echo   Alpha2        584
#> 6      E92000001 Sheffield Central E14000919   Alpha   Alpha2        969
#> 7      E92000001 Sheffield Central E14000919   Bravo   Alpha2        978
#> 8      E92000001 Sheffield Central E14000919 Charlie   Alpha2        658
#> 9      E92000001 Sheffield Central E14000919   Delta   Alpha2        551
#> 10     E92000001 Sheffield Central E14000919    Echo   Alpha2        300
#> 11     E92000001 Sheffield Central E14000919   Alpha   Alpha2        296
#> 12     E92000001 Sheffield Central E14000919   Bravo   Alpha2        616
#> 13     E92000001 Sheffield Central E14000919 Charlie   Alpha2        866
#> 14     E92000001 Sheffield Central E14000919   Delta   Alpha2        635
#> 15     E92000001 Sheffield Central E14000919    Echo   Alpha2        946
#> 16     E92000001 Sheffield Central E14000919   Alpha   Alpha2        176
#> 17     E92000001 Sheffield Central E14000919   Bravo   Alpha2        997
#> 18     E92000001 Sheffield Central E14000919 Charlie   Alpha2        545
#> 19     E92000001 Sheffield Central E14000919   Delta   Alpha2        666
#> 20     E92000001 Sheffield Central E14000919    Echo   Alpha2        526
#> 21     E92000001 Sheffield Central E14000919   Alpha   Alpha2        161
#> 22     E92000001 Sheffield Central E14000919   Bravo   Alpha2        654
#> 23     E92000001 Sheffield Central E14000919 Charlie   Alpha2        872
#> 24     E92000001 Sheffield Central E14000919   Delta   Alpha2        568
#> 25     E92000001 Sheffield Central E14000919    Echo   Alpha2        399
#> 26     E92000001 Sheffield Central E14000919   Alpha   Alpha2        645
#> 27     E92000001 Sheffield Central E14000919   Bravo   Alpha2        169
#> 28     E92000001 Sheffield Central E14000919 Charlie   Alpha2        546
#> 29     E92000001 Sheffield Central E14000919   Delta   Alpha2        359
#> 30     E92000001 Sheffield Central E14000919    Echo   Alpha2        851
#> 31     E92000001 Sheffield Central E14000919   Alpha   Bravo2        773
#> 32     E92000001 Sheffield Central E14000919   Bravo   Bravo2        774
#> 33     E92000001 Sheffield Central E14000919 Charlie   Bravo2        316
#> 34     E92000001 Sheffield Central E14000919   Delta   Bravo2        825
#> 35     E92000001 Sheffield Central E14000919    Echo   Bravo2        654
#> 36     E92000001 Sheffield Central E14000919   Alpha   Bravo2        643
#> 37     E92000001 Sheffield Central E14000919   Bravo   Bravo2        169
#> 38     E92000001 Sheffield Central E14000919 Charlie   Bravo2        364
#> 39     E92000001 Sheffield Central E14000919   Delta   Bravo2        249
#> 40     E92000001 Sheffield Central E14000919    Echo   Bravo2        365
#> 41     E92000001 Sheffield Central E14000919   Alpha   Bravo2        515
#> 42     E92000001 Sheffield Central E14000919   Bravo   Bravo2        216
#> 43     E92000001 Sheffield Central E14000919 Charlie   Bravo2        628
#> 44     E92000001 Sheffield Central E14000919   Delta   Bravo2        279
#> 45     E92000001 Sheffield Central E14000919    Echo   Bravo2        283
#> 46     E92000001 Sheffield Central E14000919   Alpha   Bravo2        121
#> 47     E92000001 Sheffield Central E14000919   Bravo   Bravo2        706
#> 48     E92000001 Sheffield Central E14000919 Charlie   Bravo2        996
#> 49     E92000001 Sheffield Central E14000919   Delta   Bravo2        169
#> 50     E92000001 Sheffield Central E14000919    Echo   Bravo2        773
#> 51     E92000001 Sheffield Central E14000919   Alpha   Bravo2        278
#> 52     E92000001 Sheffield Central E14000919   Bravo   Bravo2        785
#> 53     E92000001 Sheffield Central E14000919 Charlie   Bravo2        932
#> 54     E92000001 Sheffield Central E14000919   Delta   Bravo2        636
#> 55     E92000001 Sheffield Central E14000919    Echo   Bravo2        594
#> 56     E92000001 Sheffield Central E14000919   Alpha   Bravo2        597
#> 57     E92000001 Sheffield Central E14000919   Bravo   Bravo2        264
#> 58     E92000001 Sheffield Central E14000919 Charlie   Bravo2        363
#> 59     E92000001 Sheffield Central E14000919   Delta   Bravo2        772
#> 60     E92000001 Sheffield Central E14000919    Echo   Bravo2        835
#> 61     E92000001 Sheffield Central E14000919   Alpha Charlie2        975
#> 62     E92000001 Sheffield Central E14000919   Bravo Charlie2        950
#> 63     E92000001 Sheffield Central E14000919 Charlie Charlie2        703
#> 64     E92000001 Sheffield Central E14000919   Delta Charlie2        895
#> 65     E92000001 Sheffield Central E14000919    Echo Charlie2        915
#> 66     E92000001 Sheffield Central E14000919   Alpha Charlie2        942
#> 67     E92000001 Sheffield Central E14000919   Bravo Charlie2        390
#> 68     E92000001 Sheffield Central E14000919 Charlie Charlie2        229
#> 69     E92000001 Sheffield Central E14000919   Delta Charlie2        954
#> 70     E92000001 Sheffield Central E14000919    Echo Charlie2        973
#> 71     E92000001 Sheffield Central E14000919   Alpha Charlie2        742
#> 72     E92000001 Sheffield Central E14000919   Bravo Charlie2        905
#> 73     E92000001 Sheffield Central E14000919 Charlie Charlie2        551
#> 74     E92000001 Sheffield Central E14000919   Delta Charlie2        845
#> 75     E92000001 Sheffield Central E14000919    Echo Charlie2        676
#> 76     E92000001 Sheffield Central E14000919   Alpha Charlie2        842
#> 77     E92000001 Sheffield Central E14000919   Bravo Charlie2        761
#> 78     E92000001 Sheffield Central E14000919 Charlie Charlie2        758
#> 79     E92000001 Sheffield Central E14000919   Delta Charlie2        869
#> 80     E92000001 Sheffield Central E14000919    Echo Charlie2        931
#> 81     E92000001 Sheffield Central E14000919   Alpha Charlie2        651
#> 82     E92000001 Sheffield Central E14000919   Bravo Charlie2        292
#> 83     E92000001 Sheffield Central E14000919 Charlie Charlie2        255
#> 84     E92000001 Sheffield Central E14000919   Delta Charlie2        554
#> 85     E92000001 Sheffield Central E14000919    Echo Charlie2        111
#> 86     E92000001 Sheffield Central E14000919   Alpha Charlie2        537
#> 87     E92000001 Sheffield Central E14000919   Bravo Charlie2        908
#> 88     E92000001 Sheffield Central E14000919 Charlie Charlie2        141
#> 89     E92000001 Sheffield Central E14000919   Delta Charlie2        677
#> 90     E92000001 Sheffield Central E14000919    Echo Charlie2        201
#> 91     E92000001 Sheffield Central E14000919   Alpha   Delta2        970
#> 92     E92000001 Sheffield Central E14000919   Bravo   Delta2        893
#> 93     E92000001 Sheffield Central E14000919 Charlie   Delta2        927
#> 94     E92000001 Sheffield Central E14000919   Delta   Delta2        601
#> 95     E92000001 Sheffield Central E14000919    Echo   Delta2        369
#> 96     E92000001 Sheffield Central E14000919   Alpha   Delta2        872
#> 97     E92000001 Sheffield Central E14000919   Bravo   Delta2        586
#> 98     E92000001 Sheffield Central E14000919 Charlie   Delta2        470
#> 99     E92000001 Sheffield Central E14000919   Delta   Delta2        852
#> 100    E92000001 Sheffield Central E14000919    Echo   Delta2        968
#> 101    E92000001 Sheffield Central E14000919   Alpha   Delta2        884
#> 102    E92000001 Sheffield Central E14000919   Bravo   Delta2        112
#> 103    E92000001 Sheffield Central E14000919 Charlie   Delta2        708
#> 104    E92000001 Sheffield Central E14000919   Delta   Delta2        516
#> 105    E92000001 Sheffield Central E14000919    Echo   Delta2        688
#> 106    E92000001 Sheffield Central E14000919   Alpha   Delta2        387
#> 107    E92000001 Sheffield Central E14000919   Bravo   Delta2        296
#> 108    E92000001 Sheffield Central E14000919 Charlie   Delta2        922
#> 109    E92000001 Sheffield Central E14000919   Delta   Delta2        622
#> 110    E92000001 Sheffield Central E14000919    Echo   Delta2        903
#> 111    E92000001 Sheffield Central E14000919   Alpha   Delta2        362
#> 112    E92000001 Sheffield Central E14000919   Bravo   Delta2        762
#> 113    E92000001 Sheffield Central E14000919 Charlie   Delta2        571
#> 114    E92000001 Sheffield Central E14000919   Delta   Delta2        645
#> 115    E92000001 Sheffield Central E14000919    Echo   Delta2        555
#> 116    E92000001 Sheffield Central E14000919   Alpha   Delta2        940
#> 117    E92000001 Sheffield Central E14000919   Bravo   Delta2        406
#> 118    E92000001 Sheffield Central E14000919 Charlie   Delta2        460
#> 119    E92000001 Sheffield Central E14000919   Delta   Delta2        777
#> 120    E92000001 Sheffield Central E14000919    Echo   Delta2        976
#> 121    E92000001 Sheffield Central E14000919   Alpha    Echo2        336
#> 122    E92000001 Sheffield Central E14000919   Bravo    Echo2        636
#> 123    E92000001 Sheffield Central E14000919 Charlie    Echo2        920
#> 124    E92000001 Sheffield Central E14000919   Delta    Echo2        262
#> 125    E92000001 Sheffield Central E14000919    Echo    Echo2        144
#> 126    E92000001 Sheffield Central E14000919   Alpha    Echo2        309
#> 127    E92000001 Sheffield Central E14000919   Bravo    Echo2        506
#> 128    E92000001 Sheffield Central E14000919 Charlie    Echo2        270
#> 129    E92000001 Sheffield Central E14000919   Delta    Echo2        103
#> 130    E92000001 Sheffield Central E14000919    Echo    Echo2        517
#> 131    E92000001 Sheffield Central E14000919   Alpha    Echo2        908
#> 132    E92000001 Sheffield Central E14000919   Bravo    Echo2        181
#> 133    E92000001 Sheffield Central E14000919 Charlie    Echo2        705
#> 134    E92000001 Sheffield Central E14000919   Delta    Echo2        303
#> 135    E92000001 Sheffield Central E14000919    Echo    Echo2        406
#> 136    E92000001 Sheffield Central E14000919   Alpha    Echo2        483
#> 137    E92000001 Sheffield Central E14000919   Bravo    Echo2        244
#> 138    E92000001 Sheffield Central E14000919 Charlie    Echo2        813
#> 139    E92000001 Sheffield Central E14000919   Delta    Echo2        847
#> 140    E92000001 Sheffield Central E14000919    Echo    Echo2        368
#> 141    E92000001 Sheffield Central E14000919   Alpha    Echo2        531
#> 142    E92000001 Sheffield Central E14000919   Bravo    Echo2        906
#> 143    E92000001 Sheffield Central E14000919 Charlie    Echo2        312
#> 144    E92000001 Sheffield Central E14000919   Delta    Echo2        583
#> 145    E92000001 Sheffield Central E14000919    Echo    Echo2        804
#> 146    E92000001 Sheffield Central E14000919   Alpha    Echo2        614
#> 147    E92000001 Sheffield Central E14000919   Bravo    Echo2        874
#> 148    E92000001 Sheffield Central E14000919 Charlie    Echo2        788
#> 149    E92000001 Sheffield Central E14000919   Delta    Echo2        297
#> 150    E92000001 Sheffield Central E14000919    Echo    Echo2        574
#>     indicator2 indicator3
#> 1          367        589
#> 2          394        446
#> 3          494        251
#> 4          141        581
#> 5          139        311
#> 6          857        424
#> 7          212        481
#> 8          591        239
#> 9          343        783
#> 10         626        634
#> 11         565        592
#> 12         911        151
#> 13         696        759
#> 14         925        103
#> 15         128        438
#> 16         941        226
#> 17         583        568
#> 18         713        761
#> 19         241        632
#> 20         933        339
#> 21         552        182
#> 22         565        351
#> 23         648        785
#> 24         463        241
#> 25         743        895
#> 26         950        695
#> 27         155        760
#> 28         407        937
#> 29         796        226
#> 30         756        106
#> 31         570        234
#> 32         642        605
#> 33         576        105
#> 34         791        488
#> 35         557        623
#> 36         123        363
#> 37         867        494
#> 38         700        131
#> 39         621        499
#> 40         940        686
#> 41         822        965
#> 42         956        557
#> 43         278        238
#> 44         640        107
#> 45         148        821
#> 46         178        569
#> 47         802        473
#> 48         890        552
#> 49         968        178
#> 50         642        296
#> 51         594        767
#> 52         409        239
#> 53         139        572
#> 54         234        384
#> 55         790        786
#> 56         191        127
#> 57         203        820
#> 58         879        196
#> 59         870        972
#> 60         499        983
#> 61         683        355
#> 62         495        693
#> 63         859        859
#> 64         421        403
#> 65         682        481
#> 66         821        579
#> 67         637        250
#> 68         931        291
#> 69         479        494
#> 70         516        637
#> 71         216        490
#> 72         317        297
#> 73         997        248
#> 74         156        652
#> 75         142        665
#> 76         388        416
#> 77         627        801
#> 78         186        738
#> 79         242        559
#> 80         974        621
#> 81         131        611
#> 82         311        173
#> 83         318        415
#> 84         676        126
#> 85         671        901
#> 86         892        269
#> 87         736        682
#> 88         237        282
#> 89         654        919
#> 90         834        143
#> 91         983        752
#> 92         985        639
#> 93         466        723
#> 94         159        474
#> 95         162        829
#> 96         541        805
#> 97         930        812
#> 98         801        141
#> 99         645        105
#> 100        861        144
#> 101        141        157
#> 102        405        837
#> 103        520        549
#> 104        118        844
#> 105        695        900
#> 106        305        116
#> 107        197        932
#> 108        968        767
#> 109        745        836
#> 110        150        924
#> 111        212        853
#> 112        267        390
#> 113        833        115
#> 114        569        450
#> 115        753        140
#> 116        611        623
#> 117        166        496
#> 118        117        849
#> 119        498        456
#> 120        191        427
#> 121        223        969
#> 122        635        798
#> 123        778        160
#> 124        957        996
#> 125        408        503
#> 126        939        781
#> 127        893        518
#> 128        338        181
#> 129        264        650
#> 130        755        113
#> 131        662        845
#> 132        455        306
#> 133        896        389
#> 134        124        632
#> 135        336        105
#> 136        186        482
#> 137        918        781
#> 138        386        672
#> 139        765        554
#> 140        347        905
#> 141        486        582
#> 142        244        647
#> 143        674        454
#> 144        388        329
#> 145        588        135
#> 146        147        557
#> 147        333        403
#> 148        974        362
#> 149        568        302
#> 150        501        595
#> 
#> $meta
#>     col_name  col_type      label indicator_grouping indicator_unit
#> 1    filter1    Filter    FILTER1                 NA             NA
#> 2    filter2    Filter    FILTER2                 NA             NA
#> 3 indicator1 Indicator INDICATOR1                 NA             NA
#> 4 indicator2 Indicator INDICATOR2                 NA             NA
#> 5 indicator3 Indicator INDICATOR3                 NA             NA
#>   indicator_dp filter_hint filter_grouping_column
#> 1           NA          NA                     NA
#> 2           NA          NA                     NA
#> 3           NA          NA                     NA
#> 4           NA          NA                     NA
#> 5           NA          NA                     NA
#>