Check col_name is completed for all rows

Ensure there are no blank cells for col_name in the metafile.

Usage

precheck_meta_col_name(meta, output = "console")

Arguments

meta

A data frame of the metadata file

output

"table", "error-only", or "console". Default is "console"

Value

a console message with feedback from the check or a single row data frame if verbose = FALSE

See also

Examples

precheck_meta_col_name(example_meta)
#> ✔ The col_name column is completed for every row in the metadata.
precheck_meta_col_name(example_meta, output = "table")
#>           check result
#> 1 meta_col_name   PASS
#>                                                           message guidance_url
#> 1 The col_name column is completed for every row in the metadata.           NA