Screen files

Run all of the checks in the eesyscreener package against your data file and associated metadata file. It parses the files and runs the checks in four stages, the function will return early if any check in a stage fails.

Usage

screen_files(
  datafilename,
  metafilename,
  dataseparator,
  metaseparator,
  data_character,
  meta_character,
  datafile,
  metafile,
  api_only = FALSE
)

Arguments

api_only

Logical value indicating whether to run only the API checks.

Value

A table containing the full results of the checks with four columns:

result of the check (PASS / FAIL / ADVISORY)

message giving feedback about the check

stage that the check belongs to

name of the check

Examples

eesyscreener::screen_files(
  eesyscreener::example_data,
  eesyscreener::example_metadata
)
#> Error: 'example_data' is not an exported object from 'namespace:eesyscreener'