EES-ily generate some beefy test files

Mostly for ease of performance benchmarking. A bit crude, only does PCons for simplicity so far.

Usage

generate_test_dfs(
  years = 3000,
  pcon_names = "Sheffield Central",
  pcon_codes = "E14000919",
  num_filters = 1,
  num_indicators = 1,
  verbose = FALSE
)

Arguments

years

A vector of 4 digit years

pcon_names

A vector of PCon names

pcon_codes

A vector of PCon codes

num_filters

The number of filters to generate, currently will give each filter 5 items

num_indicators

The number of indicators to generate, currently generates random numbers between 100 and 1000

verbose

Optional console messages highlighting the number of rows being created

Value

A list containing two data frames, one data and associated meta

Examples

generate_test_dfs(2010:2015, "Sheffield Central", "E14000919", 2, 3)
#> $data
#>     time_period time_identifier           geographic_level country_name
#> 1          2010   Calendar year Parliamentary constituency      England
#> 2          2011   Calendar year Parliamentary constituency      England
#> 3          2012   Calendar year Parliamentary constituency      England
#> 4          2013   Calendar year Parliamentary constituency      England
#> 5          2014   Calendar year Parliamentary constituency      England
#> 6          2015   Calendar year Parliamentary constituency      England
#> 7          2010   Calendar year Parliamentary constituency      England
#> 8          2011   Calendar year Parliamentary constituency      England
#> 9          2012   Calendar year Parliamentary constituency      England
#> 10         2013   Calendar year Parliamentary constituency      England
#> 11         2014   Calendar year Parliamentary constituency      England
#> 12         2015   Calendar year Parliamentary constituency      England
#> 13         2010   Calendar year Parliamentary constituency      England
#> 14         2011   Calendar year Parliamentary constituency      England
#> 15         2012   Calendar year Parliamentary constituency      England
#> 16         2013   Calendar year Parliamentary constituency      England
#> 17         2014   Calendar year Parliamentary constituency      England
#> 18         2015   Calendar year Parliamentary constituency      England
#> 19         2010   Calendar year Parliamentary constituency      England
#> 20         2011   Calendar year Parliamentary constituency      England
#> 21         2012   Calendar year Parliamentary constituency      England
#> 22         2013   Calendar year Parliamentary constituency      England
#> 23         2014   Calendar year Parliamentary constituency      England
#> 24         2015   Calendar year Parliamentary constituency      England
#> 25         2010   Calendar year Parliamentary constituency      England
#> 26         2011   Calendar year Parliamentary constituency      England
#> 27         2012   Calendar year Parliamentary constituency      England
#> 28         2013   Calendar year Parliamentary constituency      England
#> 29         2014   Calendar year Parliamentary constituency      England
#> 30         2015   Calendar year Parliamentary constituency      England
#> 31         2010   Calendar year Parliamentary constituency      England
#> 32         2011   Calendar year Parliamentary constituency      England
#> 33         2012   Calendar year Parliamentary constituency      England
#> 34         2013   Calendar year Parliamentary constituency      England
#> 35         2014   Calendar year Parliamentary constituency      England
#> 36         2015   Calendar year Parliamentary constituency      England
#> 37         2010   Calendar year Parliamentary constituency      England
#> 38         2011   Calendar year Parliamentary constituency      England
#> 39         2012   Calendar year Parliamentary constituency      England
#> 40         2013   Calendar year Parliamentary constituency      England
#> 41         2014   Calendar year Parliamentary constituency      England
#> 42         2015   Calendar year Parliamentary constituency      England
#> 43         2010   Calendar year Parliamentary constituency      England
#> 44         2011   Calendar year Parliamentary constituency      England
#> 45         2012   Calendar year Parliamentary constituency      England
#> 46         2013   Calendar year Parliamentary constituency      England
#> 47         2014   Calendar year Parliamentary constituency      England
#> 48         2015   Calendar year Parliamentary constituency      England
#> 49         2010   Calendar year Parliamentary constituency      England
#> 50         2011   Calendar year Parliamentary constituency      England
#> 51         2012   Calendar year Parliamentary constituency      England
#> 52         2013   Calendar year Parliamentary constituency      England
#> 53         2014   Calendar year Parliamentary constituency      England
#> 54         2015   Calendar year Parliamentary constituency      England
#> 55         2010   Calendar year Parliamentary constituency      England
#> 56         2011   Calendar year Parliamentary constituency      England
#> 57         2012   Calendar year Parliamentary constituency      England
#> 58         2013   Calendar year Parliamentary constituency      England
#> 59         2014   Calendar year Parliamentary constituency      England
#> 60         2015   Calendar year Parliamentary constituency      England
#> 61         2010   Calendar year Parliamentary constituency      England
#> 62         2011   Calendar year Parliamentary constituency      England
#> 63         2012   Calendar year Parliamentary constituency      England
#> 64         2013   Calendar year Parliamentary constituency      England
#> 65         2014   Calendar year Parliamentary constituency      England
#> 66         2015   Calendar year Parliamentary constituency      England
#> 67         2010   Calendar year Parliamentary constituency      England
#> 68         2011   Calendar year Parliamentary constituency      England
#> 69         2012   Calendar year Parliamentary constituency      England
#> 70         2013   Calendar year Parliamentary constituency      England
#> 71         2014   Calendar year Parliamentary constituency      England
#> 72         2015   Calendar year Parliamentary constituency      England
#> 73         2010   Calendar year Parliamentary constituency      England
#> 74         2011   Calendar year Parliamentary constituency      England
#> 75         2012   Calendar year Parliamentary constituency      England
#> 76         2013   Calendar year Parliamentary constituency      England
#> 77         2014   Calendar year Parliamentary constituency      England
#> 78         2015   Calendar year Parliamentary constituency      England
#> 79         2010   Calendar year Parliamentary constituency      England
#> 80         2011   Calendar year Parliamentary constituency      England
#> 81         2012   Calendar year Parliamentary constituency      England
#> 82         2013   Calendar year Parliamentary constituency      England
#> 83         2014   Calendar year Parliamentary constituency      England
#> 84         2015   Calendar year Parliamentary constituency      England
#> 85         2010   Calendar year Parliamentary constituency      England
#> 86         2011   Calendar year Parliamentary constituency      England
#> 87         2012   Calendar year Parliamentary constituency      England
#> 88         2013   Calendar year Parliamentary constituency      England
#> 89         2014   Calendar year Parliamentary constituency      England
#> 90         2015   Calendar year Parliamentary constituency      England
#> 91         2010   Calendar year Parliamentary constituency      England
#> 92         2011   Calendar year Parliamentary constituency      England
#> 93         2012   Calendar year Parliamentary constituency      England
#> 94         2013   Calendar year Parliamentary constituency      England
#> 95         2014   Calendar year Parliamentary constituency      England
#> 96         2015   Calendar year Parliamentary constituency      England
#> 97         2010   Calendar year Parliamentary constituency      England
#> 98         2011   Calendar year Parliamentary constituency      England
#> 99         2012   Calendar year Parliamentary constituency      England
#> 100        2013   Calendar year Parliamentary constituency      England
#> 101        2014   Calendar year Parliamentary constituency      England
#> 102        2015   Calendar year Parliamentary constituency      England
#> 103        2010   Calendar year Parliamentary constituency      England
#> 104        2011   Calendar year Parliamentary constituency      England
#> 105        2012   Calendar year Parliamentary constituency      England
#> 106        2013   Calendar year Parliamentary constituency      England
#> 107        2014   Calendar year Parliamentary constituency      England
#> 108        2015   Calendar year Parliamentary constituency      England
#> 109        2010   Calendar year Parliamentary constituency      England
#> 110        2011   Calendar year Parliamentary constituency      England
#> 111        2012   Calendar year Parliamentary constituency      England
#> 112        2013   Calendar year Parliamentary constituency      England
#> 113        2014   Calendar year Parliamentary constituency      England
#> 114        2015   Calendar year Parliamentary constituency      England
#> 115        2010   Calendar year Parliamentary constituency      England
#> 116        2011   Calendar year Parliamentary constituency      England
#> 117        2012   Calendar year Parliamentary constituency      England
#> 118        2013   Calendar year Parliamentary constituency      England
#> 119        2014   Calendar year Parliamentary constituency      England
#> 120        2015   Calendar year Parliamentary constituency      England
#> 121        2010   Calendar year Parliamentary constituency      England
#> 122        2011   Calendar year Parliamentary constituency      England
#> 123        2012   Calendar year Parliamentary constituency      England
#> 124        2013   Calendar year Parliamentary constituency      England
#> 125        2014   Calendar year Parliamentary constituency      England
#> 126        2015   Calendar year Parliamentary constituency      England
#> 127        2010   Calendar year Parliamentary constituency      England
#> 128        2011   Calendar year Parliamentary constituency      England
#> 129        2012   Calendar year Parliamentary constituency      England
#> 130        2013   Calendar year Parliamentary constituency      England
#> 131        2014   Calendar year Parliamentary constituency      England
#> 132        2015   Calendar year Parliamentary constituency      England
#> 133        2010   Calendar year Parliamentary constituency      England
#> 134        2011   Calendar year Parliamentary constituency      England
#> 135        2012   Calendar year Parliamentary constituency      England
#> 136        2013   Calendar year Parliamentary constituency      England
#> 137        2014   Calendar year Parliamentary constituency      England
#> 138        2015   Calendar year Parliamentary constituency      England
#> 139        2010   Calendar year Parliamentary constituency      England
#> 140        2011   Calendar year Parliamentary constituency      England
#> 141        2012   Calendar year Parliamentary constituency      England
#> 142        2013   Calendar year Parliamentary constituency      England
#> 143        2014   Calendar year Parliamentary constituency      England
#> 144        2015   Calendar year Parliamentary constituency      England
#> 145        2010   Calendar year Parliamentary constituency      England
#> 146        2011   Calendar year Parliamentary constituency      England
#> 147        2012   Calendar year Parliamentary constituency      England
#> 148        2013   Calendar year Parliamentary constituency      England
#> 149        2014   Calendar year Parliamentary constituency      England
#> 150        2015   Calendar year Parliamentary constituency      England
#>     country_code         pcon_name pcon_code filter1 filter2 indicator1
#> 1      E92000001 Sheffield Central E14000919   Alpha   Alpha        296
#> 2      E92000001 Sheffield Central E14000919   Alpha   Alpha        616
#> 3      E92000001 Sheffield Central E14000919   Alpha   Alpha        866
#> 4      E92000001 Sheffield Central E14000919   Alpha   Alpha        635
#> 5      E92000001 Sheffield Central E14000919   Alpha   Alpha        946
#> 6      E92000001 Sheffield Central E14000919   Alpha   Alpha        176
#> 7      E92000001 Sheffield Central E14000919   Bravo   Alpha        997
#> 8      E92000001 Sheffield Central E14000919   Bravo   Alpha        545
#> 9      E92000001 Sheffield Central E14000919   Bravo   Alpha        666
#> 10     E92000001 Sheffield Central E14000919   Bravo   Alpha        526
#> 11     E92000001 Sheffield Central E14000919   Bravo   Alpha        161
#> 12     E92000001 Sheffield Central E14000919   Bravo   Alpha        654
#> 13     E92000001 Sheffield Central E14000919 Charlie   Alpha        872
#> 14     E92000001 Sheffield Central E14000919 Charlie   Alpha        568
#> 15     E92000001 Sheffield Central E14000919 Charlie   Alpha        399
#> 16     E92000001 Sheffield Central E14000919 Charlie   Alpha        645
#> 17     E92000001 Sheffield Central E14000919 Charlie   Alpha        169
#> 18     E92000001 Sheffield Central E14000919 Charlie   Alpha        546
#> 19     E92000001 Sheffield Central E14000919   Delta   Alpha        359
#> 20     E92000001 Sheffield Central E14000919   Delta   Alpha        851
#> 21     E92000001 Sheffield Central E14000919   Delta   Alpha        773
#> 22     E92000001 Sheffield Central E14000919   Delta   Alpha        774
#> 23     E92000001 Sheffield Central E14000919   Delta   Alpha        316
#> 24     E92000001 Sheffield Central E14000919   Delta   Alpha        825
#> 25     E92000001 Sheffield Central E14000919    Echo   Alpha        654
#> 26     E92000001 Sheffield Central E14000919    Echo   Alpha        643
#> 27     E92000001 Sheffield Central E14000919    Echo   Alpha        169
#> 28     E92000001 Sheffield Central E14000919    Echo   Alpha        364
#> 29     E92000001 Sheffield Central E14000919    Echo   Alpha        249
#> 30     E92000001 Sheffield Central E14000919    Echo   Alpha        365
#> 31     E92000001 Sheffield Central E14000919   Alpha   Bravo        515
#> 32     E92000001 Sheffield Central E14000919   Alpha   Bravo        216
#> 33     E92000001 Sheffield Central E14000919   Alpha   Bravo        628
#> 34     E92000001 Sheffield Central E14000919   Alpha   Bravo        279
#> 35     E92000001 Sheffield Central E14000919   Alpha   Bravo        283
#> 36     E92000001 Sheffield Central E14000919   Alpha   Bravo        121
#> 37     E92000001 Sheffield Central E14000919   Bravo   Bravo        706
#> 38     E92000001 Sheffield Central E14000919   Bravo   Bravo        996
#> 39     E92000001 Sheffield Central E14000919   Bravo   Bravo        169
#> 40     E92000001 Sheffield Central E14000919   Bravo   Bravo        773
#> 41     E92000001 Sheffield Central E14000919   Bravo   Bravo        278
#> 42     E92000001 Sheffield Central E14000919   Bravo   Bravo        785
#> 43     E92000001 Sheffield Central E14000919 Charlie   Bravo        932
#> 44     E92000001 Sheffield Central E14000919 Charlie   Bravo        636
#> 45     E92000001 Sheffield Central E14000919 Charlie   Bravo        594
#> 46     E92000001 Sheffield Central E14000919 Charlie   Bravo        597
#> 47     E92000001 Sheffield Central E14000919 Charlie   Bravo        264
#> 48     E92000001 Sheffield Central E14000919 Charlie   Bravo        363
#> 49     E92000001 Sheffield Central E14000919   Delta   Bravo        772
#> 50     E92000001 Sheffield Central E14000919   Delta   Bravo        835
#> 51     E92000001 Sheffield Central E14000919   Delta   Bravo        975
#> 52     E92000001 Sheffield Central E14000919   Delta   Bravo        950
#> 53     E92000001 Sheffield Central E14000919   Delta   Bravo        703
#> 54     E92000001 Sheffield Central E14000919   Delta   Bravo        895
#> 55     E92000001 Sheffield Central E14000919    Echo   Bravo        915
#> 56     E92000001 Sheffield Central E14000919    Echo   Bravo        942
#> 57     E92000001 Sheffield Central E14000919    Echo   Bravo        390
#> 58     E92000001 Sheffield Central E14000919    Echo   Bravo        229
#> 59     E92000001 Sheffield Central E14000919    Echo   Bravo        954
#> 60     E92000001 Sheffield Central E14000919    Echo   Bravo        973
#> 61     E92000001 Sheffield Central E14000919   Alpha Charlie        742
#> 62     E92000001 Sheffield Central E14000919   Alpha Charlie        905
#> 63     E92000001 Sheffield Central E14000919   Alpha Charlie        551
#> 64     E92000001 Sheffield Central E14000919   Alpha Charlie        845
#> 65     E92000001 Sheffield Central E14000919   Alpha Charlie        676
#> 66     E92000001 Sheffield Central E14000919   Alpha Charlie        842
#> 67     E92000001 Sheffield Central E14000919   Bravo Charlie        761
#> 68     E92000001 Sheffield Central E14000919   Bravo Charlie        758
#> 69     E92000001 Sheffield Central E14000919   Bravo Charlie        869
#> 70     E92000001 Sheffield Central E14000919   Bravo Charlie        931
#> 71     E92000001 Sheffield Central E14000919   Bravo Charlie        651
#> 72     E92000001 Sheffield Central E14000919   Bravo Charlie        292
#> 73     E92000001 Sheffield Central E14000919 Charlie Charlie        255
#> 74     E92000001 Sheffield Central E14000919 Charlie Charlie        554
#> 75     E92000001 Sheffield Central E14000919 Charlie Charlie        111
#> 76     E92000001 Sheffield Central E14000919 Charlie Charlie        537
#> 77     E92000001 Sheffield Central E14000919 Charlie Charlie        908
#> 78     E92000001 Sheffield Central E14000919 Charlie Charlie        141
#> 79     E92000001 Sheffield Central E14000919   Delta Charlie        677
#> 80     E92000001 Sheffield Central E14000919   Delta Charlie        201
#> 81     E92000001 Sheffield Central E14000919   Delta Charlie        970
#> 82     E92000001 Sheffield Central E14000919   Delta Charlie        893
#> 83     E92000001 Sheffield Central E14000919   Delta Charlie        927
#> 84     E92000001 Sheffield Central E14000919   Delta Charlie        601
#> 85     E92000001 Sheffield Central E14000919    Echo Charlie        369
#> 86     E92000001 Sheffield Central E14000919    Echo Charlie        872
#> 87     E92000001 Sheffield Central E14000919    Echo Charlie        586
#> 88     E92000001 Sheffield Central E14000919    Echo Charlie        470
#> 89     E92000001 Sheffield Central E14000919    Echo Charlie        852
#> 90     E92000001 Sheffield Central E14000919    Echo Charlie        968
#> 91     E92000001 Sheffield Central E14000919   Alpha   Delta        884
#> 92     E92000001 Sheffield Central E14000919   Alpha   Delta        112
#> 93     E92000001 Sheffield Central E14000919   Alpha   Delta        708
#> 94     E92000001 Sheffield Central E14000919   Alpha   Delta        516
#> 95     E92000001 Sheffield Central E14000919   Alpha   Delta        688
#> 96     E92000001 Sheffield Central E14000919   Alpha   Delta        387
#> 97     E92000001 Sheffield Central E14000919   Bravo   Delta        296
#> 98     E92000001 Sheffield Central E14000919   Bravo   Delta        922
#> 99     E92000001 Sheffield Central E14000919   Bravo   Delta        622
#> 100    E92000001 Sheffield Central E14000919   Bravo   Delta        903
#> 101    E92000001 Sheffield Central E14000919   Bravo   Delta        362
#> 102    E92000001 Sheffield Central E14000919   Bravo   Delta        762
#> 103    E92000001 Sheffield Central E14000919 Charlie   Delta        571
#> 104    E92000001 Sheffield Central E14000919 Charlie   Delta        645
#> 105    E92000001 Sheffield Central E14000919 Charlie   Delta        555
#> 106    E92000001 Sheffield Central E14000919 Charlie   Delta        940
#> 107    E92000001 Sheffield Central E14000919 Charlie   Delta        406
#> 108    E92000001 Sheffield Central E14000919 Charlie   Delta        460
#> 109    E92000001 Sheffield Central E14000919   Delta   Delta        777
#> 110    E92000001 Sheffield Central E14000919   Delta   Delta        976
#> 111    E92000001 Sheffield Central E14000919   Delta   Delta        336
#> 112    E92000001 Sheffield Central E14000919   Delta   Delta        636
#> 113    E92000001 Sheffield Central E14000919   Delta   Delta        920
#> 114    E92000001 Sheffield Central E14000919   Delta   Delta        262
#> 115    E92000001 Sheffield Central E14000919    Echo   Delta        144
#> 116    E92000001 Sheffield Central E14000919    Echo   Delta        309
#> 117    E92000001 Sheffield Central E14000919    Echo   Delta        506
#> 118    E92000001 Sheffield Central E14000919    Echo   Delta        270
#> 119    E92000001 Sheffield Central E14000919    Echo   Delta        103
#> 120    E92000001 Sheffield Central E14000919    Echo   Delta        517
#> 121    E92000001 Sheffield Central E14000919   Alpha    Echo        908
#> 122    E92000001 Sheffield Central E14000919   Alpha    Echo        181
#> 123    E92000001 Sheffield Central E14000919   Alpha    Echo        705
#> 124    E92000001 Sheffield Central E14000919   Alpha    Echo        303
#> 125    E92000001 Sheffield Central E14000919   Alpha    Echo        406
#> 126    E92000001 Sheffield Central E14000919   Alpha    Echo        483
#> 127    E92000001 Sheffield Central E14000919   Bravo    Echo        244
#> 128    E92000001 Sheffield Central E14000919   Bravo    Echo        813
#> 129    E92000001 Sheffield Central E14000919   Bravo    Echo        847
#> 130    E92000001 Sheffield Central E14000919   Bravo    Echo        368
#> 131    E92000001 Sheffield Central E14000919   Bravo    Echo        531
#> 132    E92000001 Sheffield Central E14000919   Bravo    Echo        906
#> 133    E92000001 Sheffield Central E14000919 Charlie    Echo        312
#> 134    E92000001 Sheffield Central E14000919 Charlie    Echo        583
#> 135    E92000001 Sheffield Central E14000919 Charlie    Echo        804
#> 136    E92000001 Sheffield Central E14000919 Charlie    Echo        614
#> 137    E92000001 Sheffield Central E14000919 Charlie    Echo        874
#> 138    E92000001 Sheffield Central E14000919 Charlie    Echo        788
#> 139    E92000001 Sheffield Central E14000919   Delta    Echo        297
#> 140    E92000001 Sheffield Central E14000919   Delta    Echo        574
#> 141    E92000001 Sheffield Central E14000919   Delta    Echo        367
#> 142    E92000001 Sheffield Central E14000919   Delta    Echo        394
#> 143    E92000001 Sheffield Central E14000919   Delta    Echo        494
#> 144    E92000001 Sheffield Central E14000919   Delta    Echo        141
#> 145    E92000001 Sheffield Central E14000919    Echo    Echo        139
#> 146    E92000001 Sheffield Central E14000919    Echo    Echo        857
#> 147    E92000001 Sheffield Central E14000919    Echo    Echo        212
#> 148    E92000001 Sheffield Central E14000919    Echo    Echo        591
#> 149    E92000001 Sheffield Central E14000919    Echo    Echo        343
#> 150    E92000001 Sheffield Central E14000919    Echo    Echo        626
#>     indicator2 indicator3
#> 1          565        592
#> 2          911        151
#> 3          696        759
#> 4          925        103
#> 5          128        438
#> 6          941        226
#> 7          583        568
#> 8          713        761
#> 9          241        632
#> 10         933        339
#> 11         552        182
#> 12         565        351
#> 13         648        785
#> 14         463        241
#> 15         743        895
#> 16         950        695
#> 17         155        760
#> 18         407        937
#> 19         796        226
#> 20         756        106
#> 21         570        234
#> 22         642        605
#> 23         576        105
#> 24         791        488
#> 25         557        623
#> 26         123        363
#> 27         867        494
#> 28         700        131
#> 29         621        499
#> 30         940        686
#> 31         822        965
#> 32         956        557
#> 33         278        238
#> 34         640        107
#> 35         148        821
#> 36         178        569
#> 37         802        473
#> 38         890        552
#> 39         968        178
#> 40         642        296
#> 41         594        767
#> 42         409        239
#> 43         139        572
#> 44         234        384
#> 45         790        786
#> 46         191        127
#> 47         203        820
#> 48         879        196
#> 49         870        972
#> 50         499        983
#> 51         683        355
#> 52         495        693
#> 53         859        859
#> 54         421        403
#> 55         682        481
#> 56         821        579
#> 57         637        250
#> 58         931        291
#> 59         479        494
#> 60         516        637
#> 61         216        490
#> 62         317        297
#> 63         997        248
#> 64         156        652
#> 65         142        665
#> 66         388        416
#> 67         627        801
#> 68         186        738
#> 69         242        559
#> 70         974        621
#> 71         131        611
#> 72         311        173
#> 73         318        415
#> 74         676        126
#> 75         671        901
#> 76         892        269
#> 77         736        682
#> 78         237        282
#> 79         654        919
#> 80         834        143
#> 81         983        752
#> 82         985        639
#> 83         466        723
#> 84         159        474
#> 85         162        829
#> 86         541        805
#> 87         930        812
#> 88         801        141
#> 89         645        105
#> 90         861        144
#> 91         141        157
#> 92         405        837
#> 93         520        549
#> 94         118        844
#> 95         695        900
#> 96         305        116
#> 97         197        932
#> 98         968        767
#> 99         745        836
#> 100        150        924
#> 101        212        853
#> 102        267        390
#> 103        833        115
#> 104        569        450
#> 105        753        140
#> 106        611        623
#> 107        166        496
#> 108        117        849
#> 109        498        456
#> 110        191        427
#> 111        223        969
#> 112        635        798
#> 113        778        160
#> 114        957        996
#> 115        408        503
#> 116        939        781
#> 117        893        518
#> 118        338        181
#> 119        264        650
#> 120        755        113
#> 121        662        845
#> 122        455        306
#> 123        896        389
#> 124        124        632
#> 125        336        105
#> 126        186        482
#> 127        918        781
#> 128        386        672
#> 129        765        554
#> 130        347        905
#> 131        486        582
#> 132        244        647
#> 133        674        454
#> 134        388        329
#> 135        588        135
#> 136        147        557
#> 137        333        403
#> 138        974        362
#> 139        568        302
#> 140        501        595
#> 141        589        560
#> 142        446        956
#> 143        251        492
#> 144        581        820
#> 145        311        926
#> 146        424        783
#> 147        481        977
#> 148        239        277
#> 149        783        854
#> 150        634        877
#> 
#> $meta
#>     col_name  col_type      label indicator_grouping indicator_unit
#> 1    filter1    Filter    FILTER1                 NA             NA
#> 2    filter2    Filter    FILTER2                 NA             NA
#> 3 indicator1 Indicator INDICATOR1                 NA             NA
#> 4 indicator2 Indicator INDICATOR2                 NA             NA
#> 5 indicator3 Indicator INDICATOR3                 NA             NA
#>   indicator_dp filter_hint filter_grouping_column
#> 1           NA          NA                     NA
#> 2           NA          NA                     NA
#> 3           NA          NA                     NA
#> 4           NA          NA                     NA
#> 5           NA          NA                     NA
#>