Run all checks against files

Run all of the checks in the eesyscreener package against your data file and associated metadata file. It takes in the paths to the CSV files then reads and parses the files and runs the checks in multiple stages, the function will return early if any check in a stage fails.

Usage

screen_csv(
  datapath,
  metapath,
  datafilename = NULL,
  metafilename = NULL,
  output = "table"
)

Arguments

datapath

Path to the data CSV file

metapath

Path to the meta CSV file

datafilename

Optional - the name of the data file, if not given it will be assumed from the path

metafilename

Optional - the name of the metadata file, if not given it will be assumed from the path

output

Control the format of output, either 'table', 'error-only', or 'console'

Value

Console messages or a list containing

A table with the full results of the checks with four columns:

result of the check (PASS / FAIL / ADVISORY)

message giving feedback about the check

group that the check belongs to

name of the check

Overall stage the checks reached

Overall message to give back to the user

Examples

# Create temp files for the example
data_file <- tempfile(fileext = ".csv")
meta_file <- tempfile(fileext = ".meta.csv")

data.table::fwrite(example_data, data_file)
data.table::fwrite(example_meta, meta_file)

screen_csv(
  data_file,
  meta_file,
  "data.csv",
  "data.meta.csv"
)
#> $results_table
#>                             check result
#> 1      check_filename_data_spaces   PASS
#> 2  check_filename_metadata_spaces   PASS
#> 3     check_filename_data_special   PASS
#> 4 check_filename_metadata_special   PASS
#> 5           check_filenames_match   PASS
#> 6                   meta_col_type   PASS
#>                                                            message guidance_url
#> 1                 'data.csv' does not have spaces in the filename.           NA
#> 2            'data.meta.csv' does not have spaces in the filename.           NA
#> 3              'data.csv' does not contain any special characters.           NA
#> 4         'data.meta.csv' does not contain any special characters.           NA
#> 5 The names of the files follow the recommended naming convention.           NA
#> 6                      col_type is always 'Filter' or 'Indicator'.           NA
#>           stage
#> 1      filename
#> 2      filename
#> 3      filename
#> 4      filename
#> 5      filename
#> 6 Precheck meta
#> 
#> $overall_stage
#> [1] "Passed"
#> 
#> $overall_message
#> [1] "Passed all checks"
#> 

# Clean up temp files
file.remove(data_file)
#> [1] TRUE
file.remove(meta_file)
#> [1] TRUE