Run all checks against data and metadata

Run all of the checks from the package against the data and metadata objects.

Usage

screen_dfs(data, meta, verbose = FALSE, stop_on_error = FALSE)

Arguments

data

data.frame, for the data table, more efficient if supplied as a lazy duckplyr data.frame

meta

data.frame, for the metadata table

verbose

logical, if TRUE prints feedback messages to console for every test, if FALSE run silently

stop_on_error

logical, if TRUE will stop with an error if the result is "FAIL", and will throw genuine warning if result is "WARNING"

Value

data.frame containing the results of the screening

Details

Provide a pair of data.frames and this will run through the checks in order.

Examples

screen_dfs(example_data, example_meta)
#>                               check result
#> 1                      col_req_meta   PASS
#> 2                  col_invalid_meta   PASS
#> 3                      col_req_data   PASS
#> 4                       col_to_rows   PASS
#> 5                     meta_col_type   PASS
#> 6                      meta_ob_unit   PASS
#> 7                     meta_col_name   PASS
#> 8                     meta_col_name   PASS
#> 9                     time_id_valid   PASS
#> 10 check_api_char_limit_column-name   PASS
#>                                                                                                       message
#> 1                                               All of the required columns are present in the metadata file.
#> 2                                                          There are no invalid columns in the metadata file.
#> 3                                                   All of the required columns are present in the data file.
#> 4  There are an equal number of rows in the metadata file (3) and non-mandatory columns in the data file (3).
#> 5                                                                 col_type is always 'Filter' or 'Indicator'.
#> 6                                             No observational units have been included in the metadata file.
#> 7                                             The col_name column is completed for every row in the metadata.
#> 8                                                    The indicator_dp column is completed for all indicators.
#> 9                                                                   The time_identifier values are all valid.
#> 10                          All filter / indicator names are less than or equal to the character limit of 50.
#>    guidance_url            stage
#> 1            NA Precheck columns
#> 2            NA Precheck columns
#> 3            NA Precheck columns
#> 4            NA Precheck columns
#> 5            NA    Precheck meta
#> 6            NA    Precheck meta
#> 7            NA    Precheck meta
#> 8            NA       Check meta
#> 9            NA    Precheck time
#> 10           NA        Check API