Run all checks against files

Run all of the checks in the eesyscreener package against your data file and associated metadata file. It parses the files and runs the checks in four stages, the function will return early if any check in a stage fails.

Usage

screen_files(datafilename, metafilename, datafile, metafile, api_only = FALSE)

Arguments

datafilename

The name of the data file.

metafilename

The name of the metadata file.

datafile

The path to the data file.

metafile

The path to the metadata file.

api_only

Logical value indicating whether to run only the API checks.

Value

A table containing the full results of the checks with four columns:

result of the check (PASS / FAIL / ADVISORY)

message giving feedback about the check

stage that the check belongs to

name of the check

Examples

screen_files(
  "data.csv",
  "data.meta.csv",
  example_data,
  example_data.meta
)
#> $results_table
#>      check                     result
#> [1,] "datafilename_spaces"     "PASS"
#> [2,] "metadatafilename_spaces" "PASS"
#>      message                                       stage
#> [1,] "The data filename does not have spaces."     1    
#> [2,] "The metadata filename does not have spaces." 1    
#> 
#> $overall_stage
#> [1] "Passed"
#> 
#> $overall_message
#> [1] "Passed all checks"
#>